chore(structure): reorganize project into envs/, scripts/single_agent/, scripts/multi_agent/, utils/, tests/
